frontend/requirements.txt
gradio
fastapi
uvicorn
langchain
langgraph
langchain-core
langchain-google-genai
python-dotenv
pandas
sqlparse
pydantic
requests==2.32.5